+++ test/features/steps/crud_database.py
@@ -36,7 +36,7 @@ def step_db_connect_test(context):
     """Send connect to database."""
     db_name = context.conf['dbname']
     context.currentdb = db_name
-    context.cli.sendline('use {0}'.format(db_name))
+    context.cli.sendline('use {0};'.format(db_name))
 
 
 @when('we connect to tmp database')