+++ changelog.md
@@ -5,6 +5,8 @@ Features:
 ---------
 * Added DSN alias name as a format specifier to the prompt (Thanks: [Georgy Frolov]).
 
+* Added DELIMITER command (Thanks: [Georgy Frolov])
+
 
 1.20.1
 ======
@@ -723,3 +725,4 @@ Bug Fixes:
 [Dick Marinus]: https://github.com/meeuw
 [François Pietka]: https://github.com/fpietka
 [Frederic Aoustin]: https://github.com/fraoustin
+[Georgy Frolov]: https://github.com/pasenor